+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * style(`Profiler`): Improve layout.
 * feat(`Profiler`): Collect object to object metadata.
 * test: In 8.2, read only classes cannot be lazy.
+* fix(`ObjectToObjectTransformer`): If target is lazy and its constructor
+  contains an eager argument, then the rest of the arguments must be eager.
 
 ## 0.7.2